Missing key exception

+++ src/AESPipe.php
@@ -3,6 +3,7 @@
 namespace Andrewhood125;
 
 use Symfony\Component\Process\Process;
+use Andrewhood125\Exceptions\MissingEncryptionKeyException;
 
 class AESPipe
 {
@@ -26,10 +27,16 @@ class AESPipe
             $out = tempnam(sys_get_temp_dir(), "");
         }
 
+        if(!getenv('ENCRYPTION_KEY')) {
+            throw new MissingEncryptionKeyException("ENCRYPTION_KEY is not set");
+        }
+
         $keyFile = getenv('ENCRYPTION_KEY');
-        $aespipe = new Process("aespipe $options -e aes256 -P ${keyFile} <${in} >${out}");
-        $aespipe->mustRun();
+
+        (new Process("aespipe $options -e aes256 -P ${keyFile} <${in} >${out}"))->mustRun();
+
         unlink($in);
+
         return $out;
     }
 }
